+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -223,6 +223,7 @@ function gwoCard() {
         "gwc_storage_and_buff",
       ];
       const expandedCards = [
+        "gwaio_combat_titans",
         "gwaio_damage_titans",
         "gwaio_enable_planetaryradar",
         "gwaio_health_titans",