+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Titan Armour Tech
   - Titan Ammunition Tech
   - Titan Engine Tech
+  - Titan Combat Tech
 
 ## v5.66.2 - 2024-07-28